ajuste da parte de contatos

+++ appMax/app/src/main/java/br/com/newton/appmax/model/dao/ContatoDao.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 class ContatoDao(
     @PrimaryKey(autoGenerate = true)
     val id: Long,
-    val idCliente: Long,
+    val idCliente: Int,
     val celular: String?, // 62988889888
     val conjuge: String?,
     val dataNascimentoConjuge: String?, // null
